Handle Comments

+++ app/Livewire/Posts/PostDashboard.php
@@ -3,6 +3,8 @@
 namespace App\Livewire\Posts;
 
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
+use Livewire\Attributes\Locked;
 use Livewire\Component;
 
 class PostDashboard extends Component
@@ -11,8 +13,12 @@ class PostDashboard extends Component
 
     public int $count;
 
+    #[Locked]
+    public bool $owned;
+
     public function delete()
     {
+        $this->authorize('delete', $this->post);
         $this->post->delete();
 
         return redirect()->to('/dashboard');
@@ -20,6 +26,7 @@ class PostDashboard extends Component
 
     public function update()
     {
+        $this->authorize('update', $this->post);
         return redirect()->to('/dashboard/edit-post/'.$this->post->id);
     }
 
@@ -30,8 +37,9 @@ class PostDashboard extends Component
 
     public function mount(Post $post)
     {
-        $this->authorize('update', $post);
+        $this->authorize('delete', $post);
         $this->post = $post;
+        $this->owned = $post->user_id == Auth::id();
         $this->count = $post->comments()->count();
     }
